Move code to Apache2 License
Copyright 2021, Xcalar Inc.

Licensed under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

    http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.

Change-Id: I37497121f918f474c82f18cda6294e37422e31d5

Former-commit-id: cbdf3c420677c86f7a2370dac658610ffdc8b338